Move "EntitiesDbSchema" file to "database/core" folder.
