+++ src/lib/database/core/EntitiesDbSchema.ts
@@ -1,7 +1,7 @@
 import type { DBSchema } from "idb";
-import type { ICoffeeBeansDB } from "./CoffeeBeansDB";
-import type { IEnhancedCoffeeBeansDB } from "./EnhancedCoffeeBeansDB";
-import type { IRecipeDB } from "./RecipeDB";
+import type { ICoffeeBeansDB } from "$lib/database/models/CoffeeBeansDB";
+import type { IEnhancedCoffeeBeansDB } from "$lib/database/models/EnhancedCoffeeBeansDB";
+import type { IRecipeDB } from "$lib/database/models/RecipeDB";
 
 export interface EntitiesDbSchema extends DBSchema {
   coffeeBeans: {
