+++ src/lib/database/manageImportExport/manageImportExport.ts
@@ -1,9 +1,9 @@
 import type { IDBPDatabase } from "idb";
 import { openEntitiesDB, COFFEEBEANS_STORE, RECIPES_STORE, ENHANCEDCOFFEEBEANS_STORE } from "$lib/database/core/core";
+import type { EntitiesDbSchema } from "$lib/database/core/EntitiesDbSchema";
 import { vacuumSoftDeletedCoffeeBeans, vacuumSoftDeletedRecipes } from "$lib/database/manageData/vacuum/vacuum";
 import { regenerateEnhancedCoffeeBeansTable } from "$lib/database/manageEnhancedCoffeeBeans";
 import { type ICoffeeBeansDB, CoffeeBeansDB, CoffeeBeansDBSubmit } from "$lib/database/models/CoffeeBeansDB";
-import type { EntitiesDbSchema } from "$lib/database/models/EntitiesDbSchema";
 import { type IRecipeDB, RecipeDB, RecipeDBSubmit } from "$lib/database/models/RecipeDB";
 import type { Count } from "$lib/domain/enhancedEntities/Count";
 import type { CoffeeBeans } from "$lib/domain/entities/CoffeeBeans";
